requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
SQLAlchemy==2.0.38
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.20
bcrypt==3.2.0
pip==24.0.0
typing_extensions==4.12.2
pydantic-settings